split student & group
