+++ src/mvc/Student.java
@@ -7,37 +7,11 @@ public class Student {
     private BigInteger id;
     private String fio;
     private Calendar date;
-    private Group group;
 
-    public class Group {
-        private BigInteger id;
-        private Integer number;
-        private String fac;
-
-        public Group(BigInteger id, Integer number, String fac) {
-            this.id = id;
-            this.number = number;
-            this.fac = fac;
-        }
-
-        public BigInteger getId() { return id; }
-
-        public void setId(BigInteger id) { this.id = id; }
-
-        public Integer getNumber() { return number; }
-
-        public void setNumber(Integer number) { this.number = number; }
-
-        public String getFac() { return fac; }
-
-        public void setFac(String fac) { this.fac = fac; }
-    }
-
-    Student(BigInteger id1, String fio, Calendar date, BigInteger id2, Integer number, String fac) {
-        this.id = id1;
+    Student(BigInteger id, String fio, Calendar date) {
+        this.id = id;
         this.fio = fio;
         this.date = date;
-        group = new Group(id2, number, fac);
     }
 
     public BigInteger getId() { return id; }
@@ -51,8 +25,6 @@ public class Student {
     public Calendar getDate() { return date; }
 
     public void setDate(Calendar date) { this.date = date; }
-
-    public Group getGroup() { return group; }
 
     @Override
     public String toString() {
